+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.16.5
+- Fixed voice chat not working when connecting to server via args